+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Server/TheNewCarlaServer.cpp
@@ -55,7 +55,8 @@ public:
 
   FPimpl(uint16_t port)
     : Server(port),
-      StreamingServer(port + 1u) {
+      StreamingServer(port + 1u)
+  {
     BindActions();
   }
 
@@ -68,6 +69,7 @@ public:
 private:
 
   void BindActions();
+
 };
 
 // =============================================================================
@@ -80,17 +82,17 @@ private:
 #  define CARLA_ENSURE_GAME_THREAD()
 #endif // WITH_EDITOR
 
-#define RESPOND_ERROR(str) { \
+#define RESPOND_ERROR(str) {                                              \
     UE_LOG(LogCarlaServer, Log, TEXT("Responding error: %s"), TEXT(str)); \
     return carla::rpc::ResponseError(str); }
 
-#define RESPOND_ERROR_FSTRING(fstr) { \
+#define RESPOND_ERROR_FSTRING(fstr) {                                 \
     UE_LOG(LogCarlaServer, Log, TEXT("Responding error: %s"), *fstr); \
     return carla::rpc::ResponseError(carla::rpc::FromFString(fstr)); }
 
 #define REQUIRE_CARLA_EPISODE() \
-    CARLA_ENSURE_GAME_THREAD(); \
-    if (Episode == nullptr) { RESPOND_ERROR("episode not ready"); }
+  CARLA_ENSURE_GAME_THREAD();   \
+  if (Episode == nullptr) { RESPOND_ERROR("episode not ready"); }
 
 // =============================================================================
 // -- Bind Actions -------------------------------------------------------------
@@ -110,13 +112,14 @@ void FTheNewCarlaServer::FPimpl::BindActions()
   {
     REQUIRE_CARLA_EPISODE();
     auto WorldObserver = Episode->GetWorldObserver();
-    if (WorldObserver == nullptr) {
+    if (WorldObserver == nullptr)
+    {
       RESPOND_ERROR("internal error: missing world observer");
     }
     return cr::EpisodeInfo{
-        Episode->GetId(),
-        cr::FromFString(Episode->GetMapName()),
-        WorldObserver->GetToken()};
+      Episode->GetId(),
+      cr::FromFString(Episode->GetMapName()),
+      WorldObserver->GetToken()};
   });
 
   Server.BindSync("get_map_info", [this]() -> R<cr::MapInfo>
@@ -125,9 +128,9 @@ void FTheNewCarlaServer::FPimpl::BindActions()
     auto FileContents = FOpenDrive::Load(Episode->GetMapName());
     const auto &SpawnPoints = Episode->GetRecommendedSpawnPoints();
     return cr::MapInfo{
-        cr::FromFString(Episode->GetMapName()),
-        cr::FromFString(FileContents),
-        MakeVectorFromTArray<cg::Transform>(SpawnPoints)};
+      cr::FromFString(Episode->GetMapName()),
+      cr::FromFString(FileContents),
+      MakeVectorFromTArray<cg::Transform>(SpawnPoints)};
   });
 
   Server.BindSync("get_actor_definitions", [this]() -> R<std::vector<cr::ActorDefinition>>
@@ -158,7 +161,7 @@ void FTheNewCarlaServer::FPimpl::BindActions()
   });
 
   Server.BindSync("set_weather_parameters", [this](
-      const cr::WeatherParameters &weather) -> R<void>
+        const cr::WeatherParameters &weather) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto *Weather = Episode->GetWeather();
@@ -171,7 +174,7 @@ void FTheNewCarlaServer::FPimpl::BindActions()
   });
 
   Server.BindSync("get_actors_by_id", [this](
-      const std::vector<FActorView::IdType> &ids) -> R<std::vector<cr::Actor>>
+        const std::vector<FActorView::IdType> &ids) -> R<std::vector<cr::Actor>>
   {
     REQUIRE_CARLA_EPISODE();
     std::vector<cr::Actor> Result;
@@ -188,8 +191,8 @@ void FTheNewCarlaServer::FPimpl::BindActions()
   });
 
   Server.BindSync("spawn_actor", [this](
-      cr::ActorDescription Description,
-      const cr::Transform &Transform) -> R<cr::Actor>
+        cr::ActorDescription Description,
+        const cr::Transform &Transform) -> R<cr::Actor>
   {
     REQUIRE_CARLA_EPISODE();
     auto Result = Episode->SpawnActorWithInfo(Transform, std::move(Description));
@@ -205,9 +208,9 @@ void FTheNewCarlaServer::FPimpl::BindActions()
   });
 
   Server.BindSync("spawn_actor_with_parent", [this](
-      cr::ActorDescription Description,
-      const cr::Transform &Transform,
-      cr::Actor Parent) -> R<cr::Actor>
+        cr::ActorDescription Description,
+        const cr::Transform &Transform,
+        cr::Actor Parent) -> R<cr::Actor>
   {
     REQUIRE_CARLA_EPISODE();
     auto Result = Episode->SpawnActorWithInfo(Transform, std::move(Description));
@@ -244,8 +247,8 @@ void FTheNewCarlaServer::FPimpl::BindActions()
   });
 
   Server.BindSync("attach_actors", [this](
-      cr::Actor Child,
-      cr::Actor Parent) -> R<void>
+        cr::Actor Child,
+        cr::Actor Parent) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto ChildView = Episode->FindActor(Child.id);
@@ -263,8 +266,8 @@ void FTheNewCarlaServer::FPimpl::BindActions()
   });
 
   Server.BindSync("set_actor_location", [this](
-      cr::Actor Actor,
-      cr::Location Location) -> R<void>
+        cr::Actor Actor,
+        cr::Location Location) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto ActorView = Episode->FindActor(Actor.id);
@@ -273,16 +276,16 @@ void FTheNewCarlaServer::FPimpl::BindActions()
       RESPOND_ERROR("unable to set actor location: actor not found");
     }
     ActorView.GetActor()->SetActorRelativeLocation(
-        Location,
-        false,
-        nullptr,
-        ETeleportType::TeleportPhysics);
+    Location,
+    false,
+    nullptr,
+    ETeleportType::TeleportPhysics);
     return R<void>::Success();
   });
 
   Server.BindSync("set_actor_transform", [this](
-      cr::Actor Actor,
-      cr::Transform Transform) -> R<void>
+        cr::Actor Actor,
+        cr::Transform Transform) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto ActorView = Episode->FindActor(Actor.id);
@@ -291,16 +294,16 @@ void FTheNewCarlaServer::FPimpl::BindActions()
       RESPOND_ERROR("unable to set actor transform: actor not found");
     }
     ActorView.GetActor()->SetActorRelativeTransform(
-        Transform,
-        false,
-        nullptr,
-        ETeleportType::TeleportPhysics);
+    Transform,
+    false,
+    nullptr,
+    ETeleportType::TeleportPhysics);
     return R<void>::Success();
   });
 
   Server.BindSync("set_actor_velocity", [this](
-      cr::Actor Actor,
-      cr::Vector3D vector) -> R<void>
+        cr::Actor Actor,
+        cr::Vector3D vector) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto ActorView = Episode->FindActor(Actor.id);
@@ -314,15 +317,15 @@ void FTheNewCarlaServer::FPimpl::BindActions()
       RESPOND_ERROR("unable to set actor velocity: not supported by actor");
     }
     RootComponent->SetPhysicsLinearVelocity(
-        vector.ToCentimeters(),
-        false,
-        "None");
+    vector.ToCentimeters(),
+    false,
+    "None");
     return R<void>::Success();
   });
 
   Server.BindSync("set_actor_angular_velocity", [this](
-      cr::Actor Actor,
-      cr::Vector3D vector) -> R<void>
+        cr::Actor Actor,
+        cr::Vector3D vector) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto ActorView = Episode->FindActor(Actor.id);
@@ -336,15 +339,15 @@ void FTheNewCarlaServer::FPimpl::BindActions()
       RESPOND_ERROR("unable to set actor angular velocity: not supported by actor");
     }
     RootComponent->SetPhysicsAngularVelocityInDegrees(
-        vector,
-        false,
-        "None");
+    vector,
+    false,
+    "None");
     return R<void>::Success();
   });
 
   Server.BindSync("add_actor_impulse", [this](
-      cr::Actor Actor,
-      cr::Vector3D vector) -> R<void>
+        cr::Actor Actor,
+        cr::Vector3D vector) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto ActorView = Episode->FindActor(Actor.id);
@@ -358,15 +361,15 @@ void FTheNewCarlaServer::FPimpl::BindActions()
       RESPOND_ERROR("unable to add actor impulse: not supported by actor");
     }
     RootComponent->AddImpulse(
-        vector.ToCentimeters(),
-        "None",
-        false);
+    vector.ToCentimeters(),
+    "None",
+    false);
     return R<void>::Success();
   });
 
   Server.BindSync("set_actor_simulate_physics", [this](
-      cr::Actor Actor,
-      bool bEnabled) -> R<void>
+        cr::Actor Actor,
+        bool bEnabled) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto ActorView = Episode->FindActor(Actor.id);
@@ -384,8 +387,8 @@ void FTheNewCarlaServer::FPimpl::BindActions()
   });
 
   Server.BindSync("apply_control_to_vehicle", [this](
-      cr::Actor Actor,
-      cr::VehicleControl Control) -> R<void>
+        cr::Actor Actor,
+        cr::VehicleControl Control) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto ActorView = Episode->FindActor(Actor.id);
@@ -403,8 +406,8 @@ void FTheNewCarlaServer::FPimpl::BindActions()
   });
 
   Server.BindSync("apply_control_to_walker", [this](
-      cr::Actor Actor,
-      cr::WalkerControl Control) -> R<void>
+        cr::Actor Actor,
+        cr::WalkerControl Control) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto ActorView = Episode->FindActor(Actor.id);
@@ -427,8 +430,8 @@ void FTheNewCarlaServer::FPimpl::BindActions()
   });
 
   Server.BindSync("set_actor_autopilot", [this](
-      cr::Actor Actor,
-      bool bEnabled) -> R<void>
+        cr::Actor Actor,
+        bool bEnabled) -> R<void>
   {
     REQUIRE_CARLA_EPISODE();
     auto ActorView = Episode->FindActor(Actor.id);
@@ -447,6 +450,101 @@ void FTheNewCarlaServer::FPimpl::BindActions()
       RESPOND_ERROR("unable to set autopilot: vehicle controller does not support autopilot");
     }
     Controller->SetAutopilot(bEnabled);
+    return R<void>::Success();
+  });
+
+  Server.BindSync("set_traffic_light_state", [this](
+        cr::Actor Actor,
+        cr::TrafficLightState trafficLightState) -> R<void>
+  {
+    REQUIRE_CARLA_EPISODE();
+    auto ActorView = Episode->GetActorRegistry().Find(Actor.id);
+    if (!ActorView.IsValid() || ActorView.GetActor()->IsPendingKill())
+    {
+      RESPOND_ERROR("unable to set state: actor not found");
+    }
+    auto TrafficLight = Cast<ATrafficLightBase>(ActorView.GetActor());
+    if (TrafficLight == nullptr)
+    {
+      RESPOND_ERROR("unable to set state: actor is not a traffic light");
+    }
+    TrafficLight->SetTrafficLightState(static_cast<ETrafficLightState>(trafficLightState));
+    return R<void>::Success();
+  });
+
+  Server.BindSync("set_traffic_light_green_time", [this](
+        cr::Actor Actor,
+        float GreenTime) -> R<void>
+  {
+    REQUIRE_CARLA_EPISODE();
+    auto ActorView = Episode->GetActorRegistry().Find(Actor.id);
+    if (!ActorView.IsValid() || ActorView.GetActor()->IsPendingKill())
+    {
+      RESPOND_ERROR("unable to set green time: actor not found");
+    }
+    auto TrafficLight = Cast<ATrafficLightBase>(ActorView.GetActor());
+    if (TrafficLight == nullptr)
+    {
+      RESPOND_ERROR("unable to set green time: actor is not a traffic light");
+    }
+    TrafficLight->SetGreenTime(GreenTime);
+    return R<void>::Success();
+  });
+
+  Server.BindSync("set_traffic_light_yellow_time", [this](
+        cr::Actor Actor,
+        float YellowTime) -> R<void>
+  {
+    REQUIRE_CARLA_EPISODE();
+    auto ActorView = Episode->GetActorRegistry().Find(Actor.id);
+    if (!ActorView.IsValid() || ActorView.GetActor()->IsPendingKill())
+    {
+      RESPOND_ERROR("unable to set yellow time: actor not found");
+    }
+    auto TrafficLight = Cast<ATrafficLightBase>(ActorView.GetActor());
+    if (TrafficLight == nullptr)
+    {
+      RESPOND_ERROR("unable to set yellow time: actor is not a traffic light");
+    }
+    TrafficLight->SetYellowTime(YellowTime);
+    return R<void>::Success();
+  });
+
+  Server.BindSync("set_traffic_light_red_time", [this](
+        cr::Actor Actor,
+        float RedTime) -> R<void>
+  {
+    REQUIRE_CARLA_EPISODE();
+    auto ActorView = Episode->GetActorRegistry().Find(Actor.id);
+    if (!ActorView.IsValid() || ActorView.GetActor()->IsPendingKill())
+    {
+      RESPOND_ERROR("unable to set red time: actor not found");
+    }
+    auto TrafficLight = Cast<ATrafficLightBase>(ActorView.GetActor());
+    if (TrafficLight == nullptr)
+    {
+      RESPOND_ERROR("unable to set red time: actor is not a traffic light");
+    }
+    TrafficLight->SetRedTime(RedTime);
+    return R<void>::Success();
+  });
+
+  Server.BindSync("freeze_traffic_light", [this](
+        cr::Actor Actor,
+        bool Freeze) -> R<void>
+  {
+    REQUIRE_CARLA_EPISODE();
+    auto ActorView = Episode->GetActorRegistry().Find(Actor.id);
+    if (!ActorView.IsValid() || ActorView.GetActor()->IsPendingKill())
+    {
+      RESPOND_ERROR("unable to alter frozen state: actor not found");
+    }
+    auto TrafficLight = Cast<ATrafficLightBase>(ActorView.GetActor());
+    if (TrafficLight == nullptr)
+    {
+      RESPOND_ERROR("unable to alter frozen state: actor is not a traffic light");
+    }
+    TrafficLight->SetTimeIsFrozen(Freeze);
     return R<void>::Success();
   });